requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.9
djangorestframework==3.12.4
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.0
